pyproject.toml
[tool.poetry]
name = "reti-tms"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
psycopg2-binary = "^2.9.3"
google-cloud-bigquery = "^2.34.1"
google-cloud-tasks = "^2.8.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
